chore(ci): Break e2e-deploy into multiple test suites
