+++ .circleci/config.yml
@@ -695,7 +695,7 @@ jobs:
       - *setup_env
       - run:
           name: "Test"
-          command: cond_spot_run_compose end-to-end 4 ./scripts/docker-compose.yml TEST=e2e_deploy_contract.test.ts
+          command: cond_spot_run_container end-to-end 4 ./src/e2e_deploy_contract/
           aztec_manifest_key: end-to-end
     <<: *defaults_e2e_test
