+++ yarn-project/end-to-end/src/e2e_token_contract/shielding.test.ts
